Fix UB of signed integer overflow
Found in Debian build and fixed by NIIBE Yutaka.

+++ src/gauche/priv/arith.h
@@ -114,16 +114,16 @@
 
 #ifndef SADDOV
 /* Portable version */
-#define SADDOV(r, v, x, y)                      \
-  do {                                          \
-    (r) = (x) + (y);                            \
-    if ((x) >= 0) {                             \
-      if ((y) >= 0 && (r) < 0) (v) = 1;         \
-      else (v) = 0;                             \
-    } else {                                    \
-      if ((y) < 0 && (r) >= 0) (v) = -1;        \
-      else (v) = 0;                             \
-    }                                           \
+#define SADDOV(r, v, x, y)                                      \
+  do {                                                          \
+    (r) = (long)((unsigned long)(x) + (unsigned long)(y));      \
+    if ((x) >= 0) {                                             \
+      if ((y) >= 0 && (r) < 0) (v) = 1;                         \
+      else (v) = 0;                                             \
+    } else {                                                    \
+      if ((y) < 0 && (r) >= 0) (v) = -1;                        \
+      else (v) = 0;                                             \
+    }                                                           \
   } while (0)
 #endif /*SADDOV*/
